pyproject.toml
[project]
name = "memdump-toolkit"
version = "1.0.0"
description = "Memory dump forensic analysis toolkit for Windows Minidumps"
authors = [{name = "Vision Security Labs"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = ["minidump", "pefile", "rapidfuzz", "rich", "click", "pyyaml"]

[project.optional-dependencies]
yara = ["yara-python"]
dotnet = ["dnfile"]
capstone = ["capstone"]
all = ["yara-python", "dnfile", "capstone"]

[project.scripts]
memdump-toolkit = "memdump_toolkit.cli:main"

[project.urls]
Repository = "https://github.com/VisionSecurityLabs/memdump-toolkit"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["memdump_toolkit"]
include = ["memdump_toolkit/rulesets.default.yml", "memdump_toolkit/signatures.default.yml"]
